+++ tests/testthat/test-dateadd.R
@@ -4,17 +4,24 @@ test_that("Date functions work on duckdb", {
   skip_if_not(eunomia_is_available())
 
   con <- DBI::dbConnect(duckdb::duckdb())
-  date_tbl <- dplyr::copy_to(con, data.frame(date1 = as.Date("1999-01-01")), name = "tmpdate", overwrite = TRUE, temporary = TRUE)
+  on.exit(DBI::dbDisconnect(con, shutdown = TRUE), add = TRUE)
+
+  date_df <- dplyr::tibble(
+    date1 = as.Date(c("2000-12-01", "2000-12-01", "2000-12-01", "2000-12-01")),
+    date2 = as.Date(c("2001-12-01", "2001-12-02", "2001-11-30", "2001-01-01"))
+  )
+
+  date_tbl <- dplyr::copy_to(con, date_df, name = "tmpdate", overwrite = TRUE, temporary = TRUE)
 
   df <- date_tbl %>%
-    dplyr::mutate(date2 = !!dateadd("date1", 1, interval = "year")) %>%
+    dplyr::mutate(date3 = !!dateadd("date1", 1, interval = "year")) %>%
     dplyr::mutate(dif_years = !!datediff("date1", "date2", interval = "year")) %>%
     dplyr::mutate(dif_days = !!datediff("date1", "date2", interval = "day")) %>%
     dplyr::collect()
 
-  expect_equal(lubridate::interval(df$date1, df$date2) / lubridate::years(1), 1)
-  expect_equal(df$dif_years, 1)
-  expect_equal(df$dif_days, 365)
+  expect_true(all((lubridate::interval(df$date1, df$date3) / lubridate::years(1)) == 1))
+  expect_equal(df$dif_years, c(1, 1, 0, 0))
+  expect_equal(df$dif_days, c(365, 366, 364,  31))
 
   df <- date_tbl %>%
     dplyr::mutate(date2 = !!dateadd("date1", 1, interval = "day")) %>%
@@ -23,10 +30,10 @@ test_that("Date functions work on duckdb", {
     dplyr::mutate(dif_days3 = !!datediff("date1", "date3", interval = "day")) %>%
     dplyr::collect()
 
-  expect_equal(lubridate::interval(df$date1, df$date2) / lubridate::days(1), 1)
-  expect_equal(lubridate::interval(df$date1, df$date3) / lubridate::days(1), -1)
-  expect_equal(df$dif_days2, 1)
-  expect_equal(df$dif_days3, -1)
+  expect_true(all(lubridate::interval(df$date1, df$date2) / lubridate::days(1) == 1))
+  expect_true(all(lubridate::interval(df$date1, df$date3) / lubridate::days(1) == -1))
+  expect_true(all(df$dif_days2 == 1))
+  expect_true(all(df$dif_days3 == -1))
 
   # can add a date and an integer column
   df <- date_tbl %>%
@@ -36,8 +43,8 @@ test_that("Date functions work on duckdb", {
     dplyr::collect()
 
 
-  expect_equal(lubridate::interval(df$date1, df$date2) / lubridate::days(1), 1)
-  expect_equal(lubridate::interval(df$date1, df$date3) / lubridate::days(1), -1)
+  expect_true(all(lubridate::interval(df$date1, df$date2) / lubridate::days(1) == 1))
+  expect_true(all(lubridate::interval(df$date1, df$date3) / lubridate::days(1) == -1))
 
 
   date_tbl2 <- dplyr::copy_to(con, data.frame(y = 2000L, m = 10L, d = 11L), name = "tmpdate2", overwrite = TRUE, temporary = TRUE)
@@ -52,9 +59,8 @@ test_that("Date functions work on duckdb", {
 
   expect_equal(as.Date(df$date_from_parts), as.Date("2000-10-11"))
 
-  DBI::dbDisconnect(con, shutdown = TRUE)
-})
 
+})
 
 test_that("Date functions work on Postgres", {
   skip_if(Sys.getenv("CDM5_POSTGRESQL_USER") == "")
@@ -112,7 +118,6 @@ test_that("Date functions work on Postgres", {
 
   DBI::dbDisconnect(con)
 })
-
 
 test_that("Date functions work on SQL Server", {
   skip_if(Sys.getenv("CDM5_SQL_SERVER_USER") == "")
@@ -178,7 +183,6 @@ test_that("Date functions work on SQL Server", {
   DBI::dbDisconnect(con)
 })
 
-
 test_that("Date functions work on Redshift", {
   skip_if(Sys.getenv("CDM5_REDSHIFT_USER") == "")
   con <- DBI::dbConnect(RPostgres::Redshift(),
@@ -194,6 +198,7 @@ test_that("Date functions work on Redshift", {
     dplyr::mutate(date2 = !!dateadd("date1", 1, interval = "year")) %>%
     dplyr::mutate(dif_years = !!datediff("date1", "date2", interval = "year")) %>%
     dplyr::mutate(dif_days = !!datediff("date1", "date2", interval = "day")) %>%
+    # dbplyr::sql_render()
     dplyr::collect()
 
   expect_equal(lubridate::interval(df$date1, df$date2) / lubridate::years(1), 1)
@@ -301,7 +306,6 @@ test_that("Date functions work on Spark", {
   DBI::dbDisconnect(con)
 })
 
-
 test_that("Date functions work on Oracle", {
   skip_if_not("OracleODBC-19" %in% odbc::odbcListDataSources()$name)
 
@@ -372,3 +376,37 @@ test_that("Date functions work on Oracle", {
   DBI::dbDisconnect(con)
 })
 
+test_that("test year, month, day functionality", {
+  con <- DBI::dbConnect(duckdb::duckdb(), ":memory:")
+  on.exit(DBI::dbDisconnect(con, shutdown = TRUE), add = TRUE)
+
+  date_tbl <- dplyr::copy_to(con,
+                             data.frame(
+                               birth_date = as.Date("1993-04-19"),
+                               date1 = as.Date("2023-04-17"),
+                               date2 = as.Date("2023-04-22")
+                              ),
+                             name = "tmp",
+                             temporary = TRUE)
+  df <- date_tbl %>%
+    dplyr::mutate(year = !!datepart("birth_date", "year")) %>%
+    dplyr::mutate(month = !!datepart("birth_date", "month")) %>%
+    dplyr::mutate(day = !!datepart("birth_date", "day")) %>%
+    dplyr::mutate(days1 = !!datediff("birth_date", "date1", "day")) %>%
+    dplyr::mutate(month1 = !!datediff("birth_date", "date1", "month")) %>%
+    dplyr::mutate(year1 = !!datediff("birth_date", "date1", "year")) %>%
+    dplyr::mutate(days2 = !!datediff("birth_date", "date2", "day")) %>%
+    dplyr::mutate(month2 = !!datediff("birth_date", "date2", "month")) %>%
+    dplyr::mutate(year2 = !!datediff("birth_date", "date2", "year")) %>%
+    dplyr::collect()
+  expect_true(df$year == 1993)
+  expect_true(df$month == 4)
+  expect_true(df$day == 19)
+  expect_true(df$days1 == 10955)
+  expect_true(df$month1 == 359)
+  expect_true(df$year1 == 29)
+  expect_true(df$days2 == 10960)
+  expect_true(df$month2 == 360)
+  expect_true(df$year2 == 30)
+
+})